requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-google-genai==1.0.3
langchain-huggingface==0.0.3
pypdf==4.2.0
faiss-cpu==1.8.0
tavily-python==0.3.3
fastapi==0.110.2
uvicorn==0.29.0
streamlit==1.33.0
python-multipart==0.0.9
requests==2.31.0
python-dotenv==1.0.1